requirements.txt
streamlit==1.34.0
langchain==0.1.17
streamlit_cropper==0.2.2
opencv-python==4.9.0.80
torchvision==0.18.0
tensorflow==2.16.1
scikit-image==0.22.0
joblib==1.4.2
pandasai==2.0.38
